+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/tiles/PedestalBlockTile.java
@@ -49,7 +49,6 @@ public class PedestalBlockTile extends ItemDisplayTile {
         //doing this here since I need crystal on server too
         ItemStack stack = getDisplayedItem();
         Item it = stack.getItem();
-        //TODO: maybe add tag
 
         if (MiscUtils.isSword(it) || stack.is(ModTags.PEDESTAL_DOWNRIGHT)) {
             this.type = DisplayType.SWORD;
@@ -92,6 +91,11 @@ public class PedestalBlockTile extends ItemDisplayTile {
         return true;
     }
 
+    @Override
+    public boolean canPlaceItemThroughFace(int index, ItemStack stack, Direction direction) {
+        return true;
+    }
+
     public enum DisplayType {
         ITEM,
         BLOCK,
@@ -107,4 +111,3 @@ public class PedestalBlockTile extends ItemDisplayTile {
     }
 
 }
-