+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/items/crafting/SoapClearRecipe.java
@@ -95,4 +95,3 @@ public class SoapClearRecipe extends CustomRecipe {
         return ModRecipes.SOAP_CLEARING.get();
     }
 }
-