+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/worldgen/CaveFilter.java
@@ -11,39 +11,47 @@ import net.minecraft.world.level.levelgen.placement.PlacementContext;
 import net.minecraft.world.level.levelgen.placement.PlacementFilter;
 import net.minecraft.world.level.levelgen.placement.PlacementModifierType;
 
+import java.util.List;
+
 public class CaveFilter extends PlacementFilter {
 
     public static final Codec<CaveFilter> CODEC =
             RecordCodecBuilder.create((instance) -> instance.group(
-                            Heightmap.Types.CODEC.fieldOf("heightmap").forGetter((p) -> p.belowHeightMap))
+                            Heightmap.Types.CODEC.listOf().fieldOf("heightmaps").forGetter((p) -> p.belowHeightMaps),
+                            Codec.BOOL.fieldOf("below_sea_level").forGetter(p -> p.belowSeaLevel)
+                    )
                     .apply(instance, CaveFilter::new));
 
-    public static class Type implements PlacementModifierType<CaveFilter>{
+    public static class Type implements PlacementModifierType<CaveFilter> {
         @Override
         public Codec<CaveFilter> codec() {
             return CODEC;
         }
     }
 
-    public static final CaveFilter BELOW_SURFACE = new CaveFilter(Heightmap.Types.MOTION_BLOCKING_NO_LEAVES);
+    private final List<Heightmap.Types> belowHeightMaps;
+    private final Boolean belowSeaLevel;
 
-
-    private final Heightmap.Types belowHeightMap;
-
-    private CaveFilter(Heightmap.Types types) {
-        this.belowHeightMap = types;
+    private CaveFilter(List<Heightmap.Types> types, Boolean belowSeaLevel) {
+        this.belowHeightMaps = types;
+        this.belowSeaLevel = belowSeaLevel;
     }
 
 
     @Override
     protected boolean shouldPlace(PlacementContext context, RandomSource random, BlockPos pos) {
         if (context.getLevel().getChunkSource() instanceof ServerChunkCache serverChunkCache) {
-            int sea = serverChunkCache.getGenerator().getSeaLevel();
-            //below sea level
             int y = pos.getY();
-            if (y > sea) return false;
-            int k = context.getHeight(this.belowHeightMap, pos.getX(), pos.getZ());
-            return y < k;
+            if(belowSeaLevel) {
+                int sea = serverChunkCache.getGenerator().getSeaLevel();
+                //below sea level
+                if (y > sea) return false;
+            }
+            for (var h : belowHeightMaps) {
+                int k = context.getHeight(h, pos.getX(), pos.getZ());
+                if (y > k) return false;
+            }
+            return true;
         }
         return false;
     }