+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/worldgen/RoadSignFeature.java
@@ -2,10 +2,15 @@ package net.mehvahdjukaar.supplementaries.common.worldgen;
 
 import com.mojang.datafixers.util.Pair;
 import com.mojang.serialization.Codec;
+import com.mojang.serialization.DataResult;
+import com.mojang.serialization.codecs.RecordCodecBuilder;
+import net.mehvahdjukaar.moonlight.api.set.wood.WoodType;
+import net.mehvahdjukaar.moonlight.api.set.wood.WoodTypeRegistry;
 import net.mehvahdjukaar.moonlight.api.util.math.MthUtils;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.CandleHolderBlock;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.NoticeBoardBlock;
 import net.mehvahdjukaar.supplementaries.common.block.blocks.WallLanternBlock;
+import net.mehvahdjukaar.supplementaries.common.block.tiles.BlockGeneratorBlockTile;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.NoticeBoardBlockTile;
 import net.mehvahdjukaar.supplementaries.common.block.tiles.SignPostBlockTile;
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
@@ -17,6 +22,7 @@ import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.nbt.StringTag;
 import net.minecraft.network.chat.Component;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.BiomeTags;
 import net.minecraft.util.Mth;
@@ -26,56 +32,138 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.WorldGenLevel;
+import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.minecraft.world.level.block.LanternBlock;
+import net.minecraft.world.level.block.SlabBlock;
 import net.minecraft.world.level.block.StairBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.AttachFace;
 import net.minecraft.world.level.levelgen.feature.Feature;
 import net.minecraft.world.level.levelgen.feature.FeaturePlaceContext;
-import net.minecraft.world.level.levelgen.feature.configurations.NoneFeatureConfiguration;
+import net.minecraft.world.level.levelgen.feature.configurations.FeatureConfiguration;
 import net.minecraft.world.level.levelgen.structure.Structure;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Function;
 
 
-public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
+public class RoadSignFeature extends Feature<RoadSignFeature.Config> {
 
-    public RoadSignFeature(Codec<NoneFeatureConfiguration> codec) {
+    public RoadSignFeature(Codec<Config> codec) {
         super(codec);
     }
 
+    public static final Codec<WoodType> WOOD_CODEC = ResourceLocation.CODEC.flatXmap(r -> {
+                WoodType w = WoodTypeRegistry.getValue(r);
+                if (w == null) return DataResult.error("No such wood type: " + r);
+                return DataResult.success(w);
+            },
+            t -> DataResult.success(t.id));
 
-    private static final BlockState COBBLE = Blocks.COBBLESTONE.defaultBlockState();
-    private static final BlockState MOSSY_COBBLE = Blocks.MOSSY_COBBLESTONE.defaultBlockState();
-    private static final BlockState FENCE = Blocks.SPRUCE_FENCE.defaultBlockState();
-    private static final BlockState WALL = Blocks.COBBLESTONE_WALL.defaultBlockState();
-    private static final BlockState MOSSY_WALL = Blocks.MOSSY_COBBLESTONE_WALL.defaultBlockState();
-    private static final BlockState TRAPDOOR = Blocks.SPRUCE_TRAPDOOR.defaultBlockState();
-    private static final BlockState LANTERN = Blocks.LANTERN.defaultBlockState().setValue(LanternBlock.HANGING, true);
-    private static final BlockState LANTERN_DOWN = Blocks.LANTERN.defaultBlockState();
-    private static final BlockState JAR = LANTERN;//ModRegistry.JAR.get().defaultBlockState(); //TODO: replace with new firefly jar
-    private static final BlockState SLAB = Blocks.SPRUCE_SLAB.defaultBlockState();
-    private static final BlockState LOG = Blocks.STRIPPED_SPRUCE_LOG.defaultBlockState();
-    private static final BlockState STONE_SLAB = Blocks.STONE_SLAB.defaultBlockState();
-    private static final BlockState STONE = Blocks.STONE.defaultBlockState();
-    private static final BlockState STAIR = Blocks.STONE_STAIRS.defaultBlockState();
+    public record Config(RandomState randomState, WoodType postWood, WoodType signWood,
+                         BlockState fence, BlockState trapdoor, BlockState slab, BlockState log,
+                         BlockState cobble, BlockState mossyCobble,
+                         BlockState wall, BlockState mossyWall,
+                         BlockState lanternUp, BlockState lanternDown,
+                         BlockState candleHolder,
+                         BlockState stone, BlockState stoneSlab,
+                         BlockState stoneStairs, String invalidMessage) implements FeatureConfiguration {
+        public static final Codec<Config> CODEC = RecordCodecBuilder.<Config>create((instance) -> instance.group(
+                RandomState.CODEC.fieldOf("random_state").forGetter(Config::randomState),
+                WOOD_CODEC.fieldOf("post_wood").forGetter(Config::postWood),
+                WOOD_CODEC.fieldOf("sign_wood").forGetter(Config::signWood),
+                BlockState.CODEC.fieldOf("cobble").forGetter(Config::cobble),
+                BlockState.CODEC.fieldOf("mossy_cobble").forGetter(Config::mossyCobble),
+                BlockState.CODEC.fieldOf("wall").forGetter(Config::wall),
+                BlockState.CODEC.fieldOf("mossy_wall").forGetter(Config::mossyWall),
+                BlockState.CODEC.fieldOf("lantern_up").forGetter(Config::lanternUp),
+                BlockState.CODEC.fieldOf("lantern_down").forGetter(Config::lanternDown),
+                BlockState.CODEC.fieldOf("candle_holder").forGetter(Config::candleHolder),
+                BlockState.CODEC.fieldOf("stone").forGetter(Config::stone),
+                BlockState.CODEC.fieldOf("stone_slab").forGetter(Config::stoneSlab),
+                BlockState.CODEC.fieldOf("stone_stairs").forGetter(Config::stoneStairs)
+        ).apply(instance, Config::of)).comapFlatMap((s) -> {
+            if (s.invalidMessage != null)
+                return DataResult.error(s.invalidMessage);
+            return DataResult.success(s);
+        }, Function.identity());
+
+        private static Config of(RandomState randomState, WoodType postWood, WoodType signWood,
+                                 BlockState cobble, BlockState mossyCobble,
+                                 BlockState wall, BlockState mossyWall,
+                                 BlockState lanternUp, BlockState lanternDown,
+                                 BlockState candleHolder,
+                                 BlockState stone, BlockState stoneSlab,
+                                 BlockState stoneStairs) {
+            String message = null;
+            Block fence = postWood.getBlockOfThis("fence");
+            if (fence == null) {
+                message = "Post wood type does not have a fence";
+                fence = Blocks.AIR;
+            }
+            Block trapdoor = postWood.getBlockOfThis("trapdoor");
+            if (trapdoor == null) {
+                message = "Post wood type does not have a trapdoor";
+                trapdoor = Blocks.AIR;
+            }
+            Block slab = postWood.getBlockOfThis("slab");
+            if (slab == null) {
+                message = "Post wood type does not have a slab";
+                slab = Blocks.AIR;
+            }
+            Block log = postWood.getBlockOfThis("stripped_log");
+            if (log == null) {
+                message = "Post wood type does not have a valid stripped log";
+                log = Blocks.AIR;
+            }
+            if (!(stoneSlab.getBlock() instanceof SlabBlock)) {
+                message = "Stone slab must be a SlabBlock, was " + stoneSlab;
+            }
+            if (!(stoneStairs.getBlock() instanceof StairBlock)) {
+                message = "Stone slab must be a StairBlock, was " + stoneStairs;
+            }
+            if (!candleHolder.hasProperty(CandleHolderBlock.FACE) || !candleHolder.hasProperty(CandleHolderBlock.LIT)) {
+                message = "Candle holder block has to have a face and lit property";
+            }
+            return new Config(randomState, postWood, signWood, fence.defaultBlockState(), trapdoor.defaultBlockState(),
+                    slab.defaultBlockState(), log.defaultBlockState(), cobble, mossyCobble, wall, mossyWall,
+                    lanternUp, lanternDown, candleHolder, stone, stoneSlab, stoneStairs, message);
+        }
+    }
+
+    private record RandomState(float doubleSignChance, float stoneChance, float stoneLanternChance,
+                               float candleHolderChance,
+                               float wallLanternChance, float doubleLanternChance, float trapdoorChance,
+                               float logChance) {
+        public static final Codec<RandomState> CODEC = RecordCodecBuilder.create((instance) -> instance.group(
+                Codec.floatRange(0, 1).fieldOf("double_sign_chance").forGetter(RandomState::doubleSignChance),
+                Codec.floatRange(0, 1).fieldOf("stone_chance").forGetter(RandomState::stoneChance),
+                Codec.floatRange(0, 1).fieldOf("stone_lantern_chance").forGetter(RandomState::stoneLanternChance),
+                Codec.floatRange(0, 1).fieldOf("candle_holder_chance").forGetter(RandomState::candleHolderChance),
+                Codec.floatRange(0, 1).fieldOf("wall_lantern_chance").forGetter(RandomState::wallLanternChance),
+                Codec.floatRange(0, 1).fieldOf("double_lantern_chance").forGetter(RandomState::doubleLanternChance),
+                Codec.floatRange(0, 1).fieldOf("trapdoor_chance").forGetter(RandomState::trapdoorChance),
+                Codec.floatRange(0, 1).fieldOf("log_chance").forGetter(RandomState::logChance)
+        ).apply(instance, RandomState::new));
+    }
+
     private static final BlockState AIR = Blocks.AIR.defaultBlockState();
     private static final BlockState PATH = Blocks.DIRT_PATH.defaultBlockState();
     private static final BlockState SANDSTONE_PATH = Blocks.SMOOTH_SANDSTONE.defaultBlockState();
-
 
     public static boolean isNotSolid(LevelAccessor world, BlockPos pos) {
         return !world.isStateAtPosition(pos, (state) -> state.isRedstoneConductor(world, pos));
     }
 
     @Override
-    public boolean place(FeaturePlaceContext<NoneFeatureConfiguration> pContext) {
+    public boolean place(FeaturePlaceContext<Config> context) {
 
-        WorldGenLevel reader = pContext.level();
-        RandomSource rand = pContext.random();
-        BlockPos pos = pContext.origin();
+
+        WorldGenLevel reader = context.level();
+        RandomSource rand = context.random();
+        BlockPos pos = context.origin();
+        Config c = context.config();
 
         /*
         if(!reader.getLevel().dimension().equals(World.OVERWORLD))return false;
@@ -124,6 +212,7 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
             }
         }
 
+        //TODO: fix mossy
         float humidity = reader.getBiome(pos).value().getDownfall();
 
 
@@ -131,14 +220,14 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
         for (int i = -2; i <= 2; ++i) {
             for (int j = -2; j <= 2; ++j) {
                 if (Math.abs(i) == 2 && Math.abs(j) == 2) continue;
-                reader.setBlock(pos.offset(i, -1, j), COBBLE, 2);
+                reader.setBlock(pos.offset(i, -1, j), c.cobble, 2);
 
                 BlockPos pathPos = pos.offset(i, 0, j);
                 double dist = pos.distToCenterSqr(pathPos.getX(), pathPos.getY(), pathPos.getZ()) / 5.2f;
 
                 if (rand.nextFloat() < dist - 0.15) continue;
                 boolean m = (humidity * 0.75) > rand.nextFloat();
-                reader.setBlock(pathPos, m ? MOSSY_COBBLE : COBBLE, 2);
+                reader.setBlock(pathPos, m ? c.mossyCobble : c.cobble, 2);
             }
         }
 
@@ -147,29 +236,32 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
         boolean m = (humidity * 0.75) > rand.nextFloat();
 
         pos = pos.above();
-        reader.setBlock(pos, m ? MOSSY_WALL : WALL, 2);
+        reader.setBlock(pos, m ? c.mossyWall : c.wall, 2);
         pos = pos.above();
-        reader.setBlock(pos, FENCE, 2);
+        reader.setBlock(pos, c.fence, 2);
         pos = pos.above();
-        reader.setBlock(pos, FENCE, 2);
+        reader.setBlock(pos, c.fence, 2);
         reader.setBlock(pos.above(), ModRegistry.BLOCK_GENERATOR.get().defaultBlockState(), 2);
-
-
+        if (reader.getBlockEntity(pos.above()) instanceof BlockGeneratorBlockTile t) {
+            t.setConfig(c);
+        }
         return true;
     }
 
 
     //post process
 
-    public static void applyPostProcess(ServerLevel level, BlockPos generatorPos, List<Pair<BlockPos, Holder<Structure>>> foundVillages) {
+    public static void applyPostProcess(Config c, ServerLevel level, BlockPos generatorPos, List<Pair<BlockPos, Holder<Structure>>> foundVillages) {
+
+
+        RandomState r = c.randomState;
+        BlockState topState = c.trapdoor;
 
         BlockPos pos = generatorPos.below(2);
 
-        BlockState topState = TRAPDOOR;
-
         List<Pair<Integer, BlockPos>> villages = new ArrayList<>();
-        for (var r : foundVillages) {
-            villages.add(Pair.of((int) Mth.sqrt((float) r.getFirst().distToCenterSqr(pos.getX(), pos.getY(), pos.getZ())), r.getFirst()));
+        for (var f : foundVillages) {
+            villages.add(Pair.of((int) Mth.sqrt((float) f.getFirst().distToCenterSqr(pos.getX(), pos.getY(), pos.getZ())), f.getFirst()));
         }
 
         //if I am in a village
@@ -178,7 +270,7 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
         if (inVillage) {
             var b = level.getBiome(pos);
             BlockState replace = b.is(BiomeTags.HAS_VILLAGE_DESERT) ? SANDSTONE_PATH : PATH;
-            replaceCobbleWithPath(level, pos, replace);
+            replaceCobbleWithPath(c, level, pos, replace);
         }
 
 
@@ -196,7 +288,7 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
 
             //only 1 sing found/ 1 tile post. always to closest village. posts that are relatively close to a village will always have two.
             //posts in a village will point away
-            if (villages.size() == 1 || (0.3 > rand.nextFloat() && villages.get(0).getFirst() > 192)) {
+            if (villages.size() == 1 || (r.doubleSignChance > rand.nextFloat() && villages.get(0).getFirst() > 192)) {
                 dist1 = villages.get(0).getFirst();
                 village1 = villages.get(0).getSecond();
                 dist2 = dist1;
@@ -213,22 +305,25 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
 
             level.setBlockAndUpdate(pos, ModRegistry.SIGN_POST.get().defaultBlockState());
             if (level.getBlockEntity(pos) instanceof SignPostBlockTile tile) {
-                tile.setHeldBlock(Blocks.SPRUCE_FENCE.defaultBlockState());
-
+                tile.setHeldBlock(c.fence);
 
                 boolean left = rand.nextBoolean();
 
 
-                var up =tile.getSignUp();
-                var down =tile.getSignDown();
+                var up = tile.getSignUp();
+                var down = tile.getSignDown();
                 up.setActive(true);
                 up.setLeft(left);
+                up.setWoodType(c.signWood);
+
                 tile.pointToward(village1, true);
 
                 down.setActive(twoSigns);
                 down.setLeft(left);
+                down.setWoodType(c.signWood);
 
                 tile.pointToward(village2, false);
+
                 if (Math.abs(up.yaw() - down.yaw()) > 90) {
                     down.toggleDirection();
                     tile.pointToward(village2, false);
@@ -267,25 +362,25 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
 
 
                 //stone
-                if (0.3 > rand.nextFloat() && Mth.degreesDifferenceAbs(tile.getPointingYaw(true) + 180, yaw) > 70) {
+                if (rand.nextFloat() < r.stoneChance && Mth.degreesDifferenceAbs(tile.getPointingYaw(true) + 180, yaw) > 70) {
                     BlockPos stonePos = pos.below().offset(backDir.getNormal());
                     if (rand.nextBoolean()) {
-                        level.setBlock(stonePos, STONE_SLAB, 2);
+                        level.setBlock(stonePos, c.stoneSlab, 2);
                     } else {
-                        level.setBlock(stonePos, STAIR.setValue(StairBlock.FACING, sideDir), 2);
+                        level.setBlock(stonePos, c.stoneStairs.setValue(StairBlock.FACING, sideDir), 2);
                     }
                     stonePos = stonePos.offset(sideDir.getNormal());
-                    level.setBlock(stonePos, STONE, 2);
-                    if (0.35 > rand.nextFloat()) {
-                        level.setBlock(stonePos.above(), hasFirefly ? JAR : LANTERN_DOWN, 3);
+                    level.setBlock(stonePos, c.stone, 2);
+                    if (rand.nextFloat() < r.stoneLanternChance) {
+                        level.setBlock(stonePos.above(), hasFirefly ? c.lanternDown : c.lanternDown, 3);
                         hasGroundLantern = true;
                     }
                     stonePos = stonePos.offset(sideDir.getNormal());
                     if (!RoadSignFeature.isNotSolid(level, stonePos.below())) {
                         if (rand.nextBoolean()) {
-                            level.setBlock(stonePos, STONE_SLAB, 2);
+                            level.setBlock(stonePos, c.stoneSlab, 2);
                         } else {
-                            level.setBlock(stonePos, STAIR.setValue(StairBlock.FACING, sideDir.getOpposite()), 2);
+                            level.setBlock(stonePos, c.stoneStairs.setValue(StairBlock.FACING, sideDir.getOpposite()), 2);
                         }
                     }
 
@@ -297,52 +392,52 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
                     //lanterns
                     pos = pos.above(2);
 
-                    BlockState light = hasFirefly ? JAR : LANTERN;
-                    if (rand.nextInt(5) == 1) {
-                        light = ModRegistry.CANDLE_HOLDERS.get(null).get().defaultBlockState()
+                    BlockState light = hasFirefly ? c.lanternUp : c.lanternUp;
+                    if (rand.nextFloat() < r.candleHolderChance) {
+                        light = c.candleHolder
                                 .setValue(CandleHolderBlock.LIT, true)
                                 .setValue(CandleHolderBlock.FACE, AttachFace.CEILING);
                     }
 
                     Direction dir = lampDir.get(rand.nextInt(lampDir.size()));
 
-                    boolean doubleSided = 0.25 > rand.nextFloat();
+                    boolean doubleSided = r.doubleLanternChance > rand.nextFloat();
                     if (doubleSided) {
                         dir = dir.getClockWise();
                     }
 
                     //wall lanterns
-                    if (0.32 > rand.nextFloat()) {
-                        topState = 0.32 > rand.nextFloat() ? TRAPDOOR : AIR;
+                    if (rand.nextFloat() < r.wallLanternChance) {
+                        topState = rand.nextFloat() < r.trapdoorChance ? c.trapdoor : AIR;
 
                         WallLanternBlock wl = ModRegistry.WALL_LANTERN.get();
-                        wl.placeOn(LANTERN_DOWN, pos.below(), dir, level);
+                        wl.placeOn(c.lanternDown, pos.below(), dir, level);
 
                         //double
                         if (doubleSided) {
-                            wl.placeOn(LANTERN_DOWN, pos.below(), dir.getOpposite(), level);
+                            wl.placeOn(c.lanternDown, pos.below(), dir.getOpposite(), level);
                         }
 
                     } else {
-                        boolean isTrapdoor = 0.4 > rand.nextFloat();
+                        boolean isTrapdoor = r.trapdoorChance > rand.nextFloat();
 
-                        if (!isTrapdoor) topState = FENCE;
+                        if (!isTrapdoor) topState = c.fence;
 
                         //double
                         if (doubleSided) {
                             BlockPos backPos = pos.relative(dir.getOpposite());
 
-                            level.setBlock(backPos, isTrapdoor ? TRAPDOOR : FENCE, 2);
+                            level.setBlock(backPos, isTrapdoor ? c.trapdoor : c.fence, 2);
 
-                            if (0.25 > rand.nextFloat()) {
-                                topState = isTrapdoor ? SLAB : LOG;
+                            if (r.logChance > rand.nextFloat()) {
+                                topState = isTrapdoor ? c.slab : c.log;
                             }
 
                             level.setBlock(backPos.below(), light, 3);
                         }
 
                         pos = pos.relative(dir);
-                        BlockState frontState = isTrapdoor ? TRAPDOOR : FENCE;
+                        BlockState frontState = isTrapdoor ? c.trapdoor : c.fence;
                         level.setBlock(pos, frontState, 2);
 
                         level.setBlock(pos.below(), light, 3);
@@ -374,7 +469,7 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
         return Component.translatable("message.supplementaries.road_sign", (((d + (s / 2)) / s) * s));
     }
 
-    private static void replaceCobbleWithPath(Level world, BlockPos pos, BlockState path) {
+    private static void replaceCobbleWithPath(Config c, Level world, BlockPos pos, BlockState path) {
         //generate cobble path
         for (int i = -2; i <= 2; ++i) {
             for (int j = -2; j <= 2; ++j) {
@@ -382,7 +477,7 @@ public class RoadSignFeature extends Feature<NoneFeatureConfiguration> {
                 if (i == 0 && j == 0) continue;
                 BlockPos pathPos = pos.offset(i, -2, j);
                 BlockState state = world.getBlockState(pathPos);
-                if (state.is(Blocks.COBBLESTONE) || state.is(Blocks.MOSSY_COBBLESTONE)) {
+                if (state.is(c.cobble.getBlock()) || state.is(c.mossyCobble.getBlock())) {
                     world.setBlock(pathPos, path, 2);
                 }
             }