+++ common/src/main/java/net/mehvahdjukaar/supplementaries/mixins/LivingEntityMixin.java
@@ -1,6 +1,5 @@
 package net.mehvahdjukaar.supplementaries.mixins;
 
-import net.mehvahdjukaar.supplementaries.common.block.blocks.RopeBlock;
 import net.mehvahdjukaar.supplementaries.configs.CommonConfigs;
 import net.mehvahdjukaar.supplementaries.reg.ModRegistry;
 import net.minecraft.util.Mth;
@@ -13,14 +12,12 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
-
-import org.jetbrains.annotations.Nullable;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin extends Entity {
@@ -40,9 +37,9 @@ public abstract class LivingEntityMixin extends Entity {
     public abstract MobEffectInstance getEffect(MobEffect pPotion);
 
     @Inject(method = "getJumpBoostPower", at = @At("RETURN"), cancellable = true)
-    private void getJumpBoostPower(CallbackInfoReturnable<Double> cir) {
+    private void getJumpBoostPower(CallbackInfoReturnable<Float> cir) {
         var effect = this.getEffect(ModRegistry.OVERENCUMBERED.get());
-        if (effect != null && effect.getAmplifier() > 0) cir.setReturnValue(cir.getReturnValue() - 0.1);
+        if (effect != null && effect.getAmplifier() > 0) cir.setReturnValue(cir.getReturnValue() - 0.1f);
     }
 
     @SuppressWarnings("ConstantConditions")
@@ -58,26 +55,9 @@ public abstract class LivingEntityMixin extends Entity {
 
                 if (this.isSuppressingSlidingDownLadder()) {
                     if (y < 0 && ((Object) this) instanceof Player) y = 0;
-                    if (ropeTicks > 0) ropeTicks--;
-                } else {
-                    if (this.level.isClientSide) {
-                        //cant use oldY since it has already been set
-                        if (this.getY() < lastY && y < -0.05) {
-                            if (RopeBlock.playEntitySlideSound(((LivingEntity) (Object) this), ropeTicks) && ropeTicks != 0) {
-                                ropeTicks = 1;
-                            } else ropeTicks++;
-                        }
-                    }
                 }
                 info.setReturnValue(new Vec3(x, y, z));
             }
-        } else if (ropeTicks > 0) ropeTicks--;
-        lastY = this.getY();
+        }
     }
-
-    @Unique
-    private int ropeTicks = 0;
-    @Unique
-    private double lastY = 0;
-
 }