+++ changelog.md
@@ -1,2 +1,7 @@
-- fixed quiver gui not closing when pausing the game
-- fixed a crash with dye block behavior
+backported some things:
+- bomb damage source
+- rope slide sound
+- suds sound
+- data driven worldgen & improved basalt feature
+- mineshaft elevator fix
+- better soap recipe